+++ students/douglas_klos/session9/mailroom/mailroom.html
@@ -12,26 +12,56 @@
             <th>Total Given</th>
             <th># of Gifts</th>
             <th>Average Gift</th>
-            <tr>
-                <td style="text-align:right">Maggie</td>
-                <td style="text-align:right">$6,000.00</td>
-                <td style="text-align:right">3</td>
-                <td style="text-align:right">$2,000.00</td>
+            <tr style="text-align:right">
+                <td>Douglas</td>
+                <td>$7,000.00</td>
+                <td>2</td>
+                <td>$3,500.00</td>
+            </tr>
+            <tr style="text-align:right">
+                <td>Maggie</td>
+                <td>$9,999.00</td>
+                <td>3</td>
+                <td>$3,333.00</td>
+            </tr>
+            <tr style="text-align:right">
+                <td>Light Yagami</td>
+                <td>$9,099.00</td>
+                <td>2</td>
+                <td>$4,549.50</td>
+            </tr>
+            <tr style="text-align:right">
+                <td>Makise Kurisu</td>
+                <td>$235,987.00</td>
+                <td>1</td>
+                <td>$235,987.00</td>
+            </tr>
+            <tr style="text-align:right">
+                <td>Youjo Senki</td>
+                <td>$24,608.00</td>
+                <td>3</td>
+                <td>$8,202.67</td>
+            </tr>
+            <tr style="text-align:right">
+                <td>Motoko Kusanagi</td>
+                <td>$385,014.00</td>
+                <td>3</td>
+                <td>$128,338.00</td>
+            </tr>
+            <tr style="text-align:right">
+                <td>Jo</td>
+                <td>$11,134.00</td>
+                <td>2</td>
+                <td>$5,567.00</td>
+            </tr>
+            <tr style="text-align:right">
+                <td>Mark</td>
+                <td>$0.00</td>
+                <td>0</td>
+                <td>$0.00</td>
             </tr>
             <tr>
-                <td style="text-align:right">Doug</td>
-                <td style="text-align:right">$16,665.00</td>
-                <td style="text-align:right">5</td>
-                <td style="text-align:right">$3,333.00</td>
-            </tr>
-            <tr>
-                <td style="text-align:right">ｷﾗ</td>
-                <td style="text-align:right">$0.00</td>
-                <td style="text-align:right">0</td>
-                <td style="text-align:right">$0.00</td>
-            </tr>
-            <tr>
-                <td colspan="4">Generated 2019-03-14 19:29:28</td>
+                <td colspan="4">Generated 2019-03-17 19:11:50</td>
             </tr>
         </table>
     </body>
